+++ lib/src/main/java/com/smileidentity/networking/NetworkingUtil.kt
@@ -10,6 +10,7 @@ import java.io.BufferedOutputStream
 import java.io.File
 import java.io.FileOutputStream
 import java.util.zip.ZipEntry
+import java.util.zip.ZipException
 import java.util.zip.ZipOutputStream
 
 fun calculateSignature(timestamp: String): String {
@@ -24,16 +25,18 @@ fun calculateSignature(timestamp: String): String {
 }
 
 fun UploadRequest.zip(): File {
+    val uploadRequest = deDupedUploadRequest(this)
     val zipFile = File.createTempFile("upload", ".zip")
     val zipOutputStream = ZipOutputStream(BufferedOutputStream(FileOutputStream(zipFile)))
 
     // Write info.json
     zipOutputStream.putNextEntry(ZipEntry("info.json"))
-    zipOutputStream.write(moshi.adapter(UploadRequest::class.java).toJson(this).toByteArray())
+    val infoJson = moshi.adapter(UploadRequest::class.java).toJson(uploadRequest)
+    zipOutputStream.write(infoJson.toByteArray())
     zipOutputStream.closeEntry()
 
     // Write images
-    this.images.forEach { imageInfo ->
+    uploadRequest.images.forEach { imageInfo ->
         zipOutputStream.putNextEntry(ZipEntry(imageInfo.image.name))
         imageInfo.image.inputStream().use { it.copyTo(zipOutputStream) }
         zipOutputStream.closeEntry()
@@ -43,6 +46,33 @@ fun UploadRequest.zip(): File {
     zipFile.deleteOnExit()
     return zipFile
 }
+
+/**
+ * This function takes an [UploadRequest] and returns a new [UploadRequest] with duplicate images
+ * copied to a new File with a new name. This is necessary because we can't have duplicate entries
+ * in a zip file.
+ *
+ * If we don't do this, then we will crash when attempting to include the file in the Zip with a
+ * [ZipException] stating that the entry already exists
+ */
+private fun deDupedUploadRequest(uploadRequest: UploadRequest) = uploadRequest.copy(
+    images = uploadRequest.images
+        .groupBy { it.image.name }
+        .flatMap { (fileName, images) ->
+            if (images.size > 1) {
+                images.mapIndexed { index, imageInfo ->
+                    val fileNameWithoutExtension = fileName.substringBeforeLast(".")
+                    val fileNameExtension = fileName.substringAfterLast(".")
+                    val newFileName = "$fileNameWithoutExtension-$index.$fileNameExtension"
+                    val newFile = File(SmileID.fileSavePath, newFileName)
+                    imageInfo.image.copyTo(newFile, overwrite = true)
+                    imageInfo.copy(image = newFile)
+                }
+            } else {
+                images
+            }
+        },
+)
 
 fun File.asSelfieImage() = UploadImageInfo(
     imageTypeId = ImageType.SelfieJpgFile,